move navigation link items to data folder
