+++ Web/starter_project_vanguard/data/layout/nav-items.ts
